Renames VideoThumnail files properly, Adds navigation to VideoWithControls
